Fixed linter problems
